Read per-chapter config from markdown front matter
This replaces trying to parse the publication date from the chapter
title with using a `pub_date` config from the article's front matter,
making the date more customizable and allowing more freedom concerning
the naming of the chapters' files.

For now the publication date and an optional description is the only
information used from a front matter.